Struct serde_url::Url

pub struct Url { /* fields omitted */ }

Opaque type that can be serialized/deserialized and acts like a string.

The goal is not URL maniplutation, but rather reading and writing URL's, as well as ensuring a consistent format.

Note Usernames/Passwords

Non-UTF8 usernames and passwords will generate an error.

Note Hashing & Equality

This type implements std::hash::Hash, it will use the output of get_string() for the purposes of hashing or comparison. Either as a utf8 string, or array of bytes.

Methods

impl Url

pub fn new<S>(input: &S) -> Result<Url, UrlFault> where
S: AsRef<str>,

new is a generally entrypoint for constructing a Url also applicable are from_str and serde::Deserialize

use serde_url::Url;

let data = "https://google.com";
let url = match Url::new(&data) {
    Err(e) => panic!("failed to parse url:{} error:{:?}", data, e),
    Ok(url) => url,
};
assert_eq!(url, "https://google.com/");

pub fn get_string<'a>(&'a self) -> &'a str

get_string just returns a string

pub fn get_input<'a>(&'a self) -> &'a str

get_input returns the input argument

pub fn get_scheme<'a>(&'a self) -> &'a str

get_scheme returns the URL's scheme

pub fn get_username<'a>(&'a self) -> Option<&'a str>

get_username returns the percentage decoded username if one is present.

pub fn get_password<'a>(&'a self) -> Option<&'a str>

get_password returns the percentage decoded password if one is present.

pub fn get_host<'a>(&'a self) -> Option<Host<&'a str>>

get_host returns host information. This maybe a domain name, or IP address. You are encouraged to inspect the value if interested.

pub fn get_port(&self) -> Option<u16>

get_port returns host information about the port.

pub fn get_origin<'a>(&'a self) -> Option<Origin<'a>>

get_origin returns an a non-opaque origin. If one is present. This contains the host and port, as well as scheme information.

pub fn get_path<'a>(&'a self) -> Option<&'a Path>

get_path returns the path component of the URL

Note

attempts to decode the percentage encoding if any is present.

pub fn get_path_str<'a>(&'a self) -> Option<&'a str>

get_path_str returns the path component of the URL, as a str vs Path, which maybe preferable in some scenarios.

Note

attempts to decode the percentage encoding if any is present.

pub fn get_query_info<'a>(&'a self) -> Option<QueryData<'a>>

get_query_info returns information about query parameters