Struct serde_url::Url

pub struct Url { /* fields omitted */ }

Opaque type that can be serialized/deserialized and acts like a string.

The goal is not URL maniplutation, but rather reading and writing URL's, as well as ensuring a consistent format.

Note Usernames/Passwords

Non-UTF8 usernames and passwords will generate an error.

Note Hashing & Equality

This type implements std::hash::Hash, it will use the output of get_string() for the purposes of hashing or comparison. Either as a utf8 string, or array of bytes.

Methods

impl Url

pub fn new<S>(input: &S) -> Result<Url, UrlFault> where
S: AsRef<str>,

new is a generally entrypoint for constructing a Url also applicable are from_str and serde::Deserialize

use serde_url::Url;

let data = "https://google.com";
let url = match Url::new(&data) {
    Err(e) => panic!("failed to parse url:{} error:{:?}", data, e),
    Ok(url) => url,
};
assert_eq!(url, "https://google.com/");

pub fn get_string<'a>(&'a self) -> &'a str

get_string returns the normalized URL representation

use serde_url::Url;

let data = "https://google.com";
let url = Url::new(&data).unwrap();
assert_eq!(url.get_string(), "https://google.com/");

This method does not require allocations.

This representation of the URL will be used for most common string tasks namely:

AsRef<str>

Deref<str>

Hash

AsRef<[u8]>

this allows for a great range of flexiability.

pub fn get_input<'a>(&'a self) -> &'a str

get_input returns the input argument

use serde_url::Url;

let data = "https://google.com";
let url = Url::new(&data).unwrap();
assert!(data == url.get_input());
assert!(data != url.get_string());

pub fn get_scheme<'a>(&'a self) -> &'a str

get_scheme returns the URL's scheme

use serde_url::Url;

let data = "https://google.com";
let url = Url::new(&data).unwrap();
assert!(url.get_scheme() == "https");

pub fn get_username<'a>(&'a self) -> Option<&'a str>

get_username returns the percentage decoded username if one is present.

use serde_url::Url;

let data = "https://google.com";
let url = Url::new(&data).unwrap();
assert!(url.get_username() == Option::None);

If we include a username/password

use serde_url::Url;

let data = "https://janedoe:hunter2@google.com";
let url = Url::new(&data).unwrap();
assert!(url.get_username() == Option::Some("janedoe"));
assert!(url.get_password() == Option::Some("hunter2"));

Escape sequences are automatically handled.

use serde_url::Url;

let data = "ftps://john%20doe@google.com";
let url = Url::new(&data).unwrap();
assert!(url.get_username() == Option::Some("john doe"));
assert!(url.get_password() == Option::None);
assert!(url.get_scheme() == "ftps");

pub fn get_password<'a>(&'a self) -> Option<&'a str>

get_password returns the password from the username field. See get_username for more information.

pub fn get_host<'a>(&'a self) -> Option<Host<&'a str>>

get_host returns host information. This maybe a domain name, or IP address. You are encouraged to inspect the value if interested.

use serde_url::{Url,Host};
use std::net::Ipv4Addr;

let data = "https://192.168.0.1:8080/";
let url = Url::new(&data).unwrap();
match url.get_host().unwrap() {
    Host::Ipv4(ipaddr) => assert!(ipaddr == Ipv4Addr::new(192,168,0,1)),
    _ => panic!("should not occur"),
};

Also works with IPV6

use serde_url::{Url,Host};
use std::net::Ipv6Addr;

let data = "https://[fe80::1]:8080/";
let url = Url::new(&data).unwrap();
match url.get_host().unwrap() {
    Host::Ipv6(ipaddr) => assert!(ipaddr == Ipv6Addr::new(0xfe80,0,0,0,0,0,0,0x01)),
    _ => panic!("should not occur"),
};

And with domains

use serde_url::{Url,Host};

let data = "https://github.com:8080/";
let url = Url::new(&data).unwrap();
match url.get_host().unwrap() {
    Host::Domain(domain) => assert!(domain == "github.com"),
    _ => panic!("should not occur"),
};

pub fn get_port(&self) -> Option<u16>

get_port returns host information about the port.

use serde_url::{Url,Host};

let data = "https://github.com:8080/";
let url = Url::new(&data).unwrap();
assert!(url.get_port().unwrap() == 8080u16);

pub fn get_origin<'a>(&'a self) -> Option<Origin<'a>>

get_origin returns an a non-opaque origin. If one is present. This contains the host and port, as well as scheme information.

pub fn get_path<'a>(&'a self) -> Option<&'a Path>

get_path returns the path component of the URL

Note

attempts to decode the percentage encoding if any is present.

pub fn get_path_str<'a>(&'a self) -> Option<&'a str>

get_path_str returns the path component of the URL, as a str vs Path, which maybe preferable in some scenarios.

Note

attempts to decode the percentage encoding if any is present.

pub fn get_query_info<'a>(&'a self) -> Option<QueryData<'a>>

get_query_info returns information about query parameters